Fixed Multitap Support
If you set second RA-controller to a multitap device the second RA-controller will control SNES-controllers from 2 to 5, while input of RA-controller 3 to 5 is being ignored (e.g. in Bomberman).
This problem is fixed here. Sending pressed buttons of second RA-controller to all SNES-controllers except first seems to be needed for games which doesn't support multitap but have a two player mode being used with a connected multitap, so these cases are distinguished in the fix.