Assign map descriptors after ROM is loaded
Moved the assignment of map.descriptors and map.num_descriptors
until after the ROM is loaded.  Since the descriptors are initialized as
part of loading the ROM, assigning these struct members prior to loading
the ROM meant that retroarch would never see any memory maps.  After
this change, I can see the maps in the retroarch log output, as
expected.